(interbit-ui-components, app-account, app-todo-list): refactor renderInput and export IbField wrapper component for redux-form's Field
