+++ packages/app-account/src/components/ProfileForm.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import { FormGroup, FormControl, ControlLabel, Button } from 'react-bootstrap'
-import { Field, reduxForm } from 'redux-form'
-import { ContentBox, validation, renderInputNew } from 'interbit-ui-components'
+import { reduxForm } from 'redux-form'
+import { ContentBox, validation, IbField } from 'interbit-ui-components'
 import PropTypes from 'prop-types'
 
 import formNames from '../constants/formNames'
@@ -69,9 +69,8 @@ class ProfileForm extends Component {
       <form onSubmit={handleSubmit}>
         <FormGroup controlId="formHorizontalEmail">
           <ControlLabel>Email</ControlLabel>
-          <Field
+          <IbField
             type="email"
-            component={renderInputNew}
             name="email"
             placeholder="Email"
             validate={[validation.email]}
@@ -80,22 +79,12 @@ class ProfileForm extends Component {
 
         <FormGroup controlId="formHorizontalAlias">
           <ControlLabel>Username</ControlLabel>
-          <Field
-            type="text"
-            component={renderInputNew}
-            name="alias"
-            placeholder="Username"
-          />
+          <IbField type="text" name="alias" placeholder="Username" />
         </FormGroup>
 
         <FormGroup controlId="formHorizontalName">
           <ControlLabel>Name</ControlLabel>
-          <Field
-            type="text"
-            component={renderInputNew}
-            name="name"
-            placeholder="Name"
-          />
+          <IbField type="text" name="name" placeholder="Name" />
         </FormGroup>
 
         <FormGroup>
