+++ packages/interbit-ui-components/src/index.js
@@ -37,6 +37,7 @@ export { default as LinkWrapper } from './components/UIKit/LinkWrapper'
 export { default as ModalWrapper } from './components/UIKit/ModalWrapper'
 
 // Redux form
-export { default as validation } from './help/newReduxForm/validation'
-export { renderInputNew } from './help/newReduxForm/renderInput'
+export { default as validation } from './components/ReduxForm/validation'
+export { renderInputNew } from './components/ReduxForm/IbField'
+export { default as IbField } from './components/ReduxForm/IbField'
 export { renderInput } from './help/reduxForm/reduxForm'
